+++ lib/utils/color_extension.dart
@@ -1,0 +1,5 @@
+extension ColorExtension on String {
+  String get red => '\x1B[31m$this\x1B[0m';
+  String get green => '\x1B[32m$this\x1B[0m';
+  String get yellow => '\x1B[33m$this\x1B[0m';
+}